Getting started
Ok, so I've mashed out some code that traverses a parse tree generated by
Esprima and does some simple tests.

I've also written a whole bunch of tests to check that various errors get
caught and that valid code gets through.

There are lots of things that don't get tested, lots of ugly code in the
implementation, lots of false positives in the style tests that are there...
but I wanted to get this committed so that I have something to rollback to
when I inevitably trash things :)