Ok, getting out of hand again
I completely rewrote the checkers to have  much cleverer whitespace parsing,
which was great. But I had still had problems.

So, this evening I started rewriting AGAIN to use a slightly modified parse
tree that includes all of the tokens for each node. I'm commiting now with
the work not finished so that I still have some evidence of the intermediate
cleverness in case I want to go back to that!

Must learn to commit more often...